refactor: 계산식에 대한 검증 및 생성하는 클래스 생성 및 calculate로직 수정
